Changed TestMacroLibrary to StandardMacroLibrary
These are macros that anyone has out the box when compiling XVNML and running the dialogue writer. Meaning we don't have to rewrite all of these functions.

Also, delay was changed to not use "Thread.Sleep" of fear of stalling Unity process for waiting of a dialogue. Instead, we used a Timer from System.Timers.