Removed Asynchronous Method: Stuck with Parallel.ForLoop
After some time testing Asynchronous methods and setting the output, we've realized that this isn't want we wanted for our code. We went in and took out all of the async methods, and had our original solution (but with the improvements from the MacroInvoker, DialogueWriterProcessor, and the DialogueWriter itself). 

We ran some test inside Unity to make sure there was no hanging, and it seems to be consistent.