Expression Changes Implemented
We made the DialogueWriterProcess able to change and trigger an event if an expression or voice of a cast has changed.

There's also an event that can be triggered if an entirely new cast has changed just by looking at the name. If the name is different, it's best to assume this is a different character or cast speaking.