Didn't consider Floats, Doubles and Booleans.
Whenever you call a macro and pass in a value, we forgot to account for Floats, Doubles, and Booleans. This made creating the method endpoints of those macros hard to implement. What also made it difficult was differentiating between an Unsigned Int and an Int. With this, we allow suffixes such a "F", "D", "L", "I", or "U" to be added at the end of the number, affecting how the value is parsed for the MacroInvoker to evaluate and confirm validity.