Patched "DialogueWriterProcessor", "SkriptrParser", and "MacroInvoker"
DialogueWriterProcessor would cause slow down from the previous implementation (which used a stack to know the return points). 

SkriptrLine data for return points was inaccurate to the model due to the logic for SkriptrParser when accommodating for Named Lines (A line with "[...]" above it).

Lastly, there was no check for the amount of arguments in a macro call and its corresponding method, which resulted in an exception. Adding the check resolved this issue, and finds the correct macro variant.