XVNMLObj is cacheable using JSON serialization
To address an issue regarding building an XVNMLObj during a runtime of an application, I implemented some caching logic where after the build, it generates a JSON file based from its name, followed by "cache.json". If this file exists when the XVNMLObj attempts to build again, it'll use this cache data instead, resulting in a much faster reading time.