Added SearchElement method
SearchElement method is different from GetElement where from the tag that the method is called from, it'll do a "deep-search" on every child element until it finds the element it's looking for (or the first occurrence of the element that satisfies the type). It's a very expensive operation, and the data that it returns should be cached after use.