Optimization of Tokenizing/Parser Process
Instead of instantiated a new Tokenizer every time, at the point of creating a tag (especially for a dialogue tag), a set of tokens are passed in for the SkriptrParser and SkriptrLine to use, utilizing the same single tokenizer that was instantiated at the start of reading the document.

For context, if you used older version of it in Unity, a very lengthy and complicated XVNML would take 27 seconds until Unity starts play. That time has been reduced between 15 and 16 seconds respectively.