Further optimized Tokenizer and Parsing Processes
By being more specific at the start of tokenization, other parses that needs to know more about the Unique Token has the option to "Complicate" it, breaking that token down to even smaller tokens.

The inclusion of Unique Tags Reduced the definedTokens inside a Tokenizer from 15K+ all the way down to ~430 tokens.

For any parser that needs to be more specific (like the Skriptr Parser), it could break it down when needed powering from the Main Tokenizer's defined tokens, reducing indexing time, resulting in an almost instant XVNML compilation.