Fixed Named Lines on first line
Whenever you have a named line as the first line of dialogue, the Parser would fail to recognize that it's in a Dialogue Tag.

The easy fixed was to check if the TagTypeName is equal to "dialogue". Once the condition is true, it'll collected Dialogue value within the <dialogue> tag, even if the first token is a "[".