Update on Macro Tags
Macro Tags can now take in multiple other macro tags as children elements and execute them.

We've also introduced a new tag called "arg" for macro symbols that take in multiple arguments.

There is now no need to flag the type of the argument, as it's inferred by the parser.