Override Feature Integrated
You can further extend pre-built XVNML tags by including the #override flag to a tag. Then, prior to building the tag, use the UserOverrideManager to include new Parameters or Flags to a specific tag type, or a tag of a particular name and type.

That way, we don't need to add requirements inside the actual tool. Instead, if you want a particular tag to do something special, and it doesn't have a flag to check it, you can use this new Override Feature to extend the tag and change its behavior beyond what it could normally do.